+++ gridkit/io.py
@@ -1,3 +1,6 @@
+import warnings
+from typing import Tuple
+
 import numpy
 import rasterio
 from pyproj import CRS, Transformer
@@ -5,18 +8,36 @@ from pyproj import CRS, Transformer
 from gridkit.rect_grid import BoundedRectGrid
 
 
-def read_geotiff(path, bands=1, bounds=None, bounds_crs=None, border_buffer=0):
+def read_raster(
+    path: str,
+    bounds: Tuple[float, float, float, float] = None,
+    bounds_crs: CRS = None,
+    border_buffer: float = 0,
+):
     """Read data from a GeoTIFF
+
+    Parameters
+    ----------
+    path: :class:`str`
+        The path to the file. This needs to be a file that is supported by rasterio.
+    bounds: Tuple(float, float, float, float)
+        The bounds of the are of interest. Only the data within the supplied bounds is read from the input file.
+    bounds_crs: `pyproj.CRS`
+        The Coordinte Reference System (CRS) of the supploed bounds.
+        If the CRS of the bounds does not match that of the input files,
+        the bounds are converted to that of the input file before reading.
+    border_buffer: :class:`int`
+        A buffer to apply to the supplied `bounds` to read in a larger slice of the area.
 
     Returns
     -------
     :class:`~gridkit.rect_grid.BoundedRectGrid`
-        The contents of the GeoTIFF read into a Raster
+        The contents of the GeoTIFF in the form of a BoundedRectGrid
 
+    See also
+    --------
+    :func:`.write_raster`
     """
-    if not isinstance(bands, int):
-        raise NotImplementedError("Reading multiple bounds is not yet supported.")
-
     with rasterio.open(path) as raster_file:
         crs = str(raster_file.crs)
 
@@ -44,14 +65,41 @@ def read_geotiff(path, bands=1, bounds=None, bounds_crs=None, border_buffer=0):
             b = raster_file.bounds if bounds is None else bounds
             bounds = (b.left, b.bottom, b.right, b.top)
 
-        data = raster_file.read(bands, window=window)
+        data = raster_file.read(1, window=window)
         nodata = raster_file.nodata
 
     grid = BoundedRectGrid(data, bounds=bounds, crs=crs, nodata_value=nodata)
     return grid
 
 
+def read_geotiff(*args, bands=1, **kwargs):
+    """Deprecated, please refer to :func:`read_raster`"""
+    warnings.warn(
+        "gridkit's 'read_geotiff' has been replaced with 'read_raster' and will be removed in a future update."
+    )
+    return read_raster(*args, **kwargs)
+
+
 def write_raster(grid, path):
+    """Write a BoundedRectGrid to a raster file (eg .tiff).
+
+    Parameters
+    ----------
+    grid: :class:`.BoundedRectGrid`
+        The grid to write to a raster file.
+        This can only be a BoundedRectGrid.
+    path: :class:`str`
+        The locatin of the file to write to (eg ./my_raster.tiff).
+
+    Returns
+    -------
+    :class:`str`
+        The path pointing to the written file
+
+    See also
+    --------
+    :func:`read_raster`
+    """
     transform = rasterio.transform.from_bounds(*grid.bounds, grid.width, grid.height)
     with rasterio.open(
         path,